Refactor table accessing from the api package to a separate package.
