+++ api/tables/doc.go
@@ -1,0 +1,10 @@
+package tables
+
+// The GUI frequently needs to display tables. This package manages
+// the GUI's table access:
+
+// 1. Central place for managing where result sets are stored and
+//    accessed for the different tables needed
+
+// 2. Apply any result set transformations needed on the underlying
+//    result set implementation.
